tests/requirements.txt
asgi-lifespan~=1.0
httpx~=0.18
pytest-asyncio>=0.15.1